add gradient clippings

+++ lib/az_data_layer/roidb.py
@@ -53,8 +53,8 @@ def prepare_roidb(imdb):
         ex_rois, zb = _compute_ex_rois(imdb.image_size(i), gt_rois)
         
         roidb[i]['zoom_gt'] = zb.astype(np.bool, copy=False)
-        roidb[i]['ex_boxes'] = ex_rois.astype(np.uint16, copy=False)
-        roidb[i]['gt_boxes'] = gt_rois.astype(np.uint16, copy=False)
+        roidb[i]['ex_boxes'] = ex_rois.astype(np.float32, copy=False)
+        roidb[i]['gt_boxes'] = gt_rois.astype(np.float32, copy=False)
         
         assert np.all(roidb[i]['ex_boxes'][:, 0] < roidb[i]['ex_boxes'][:, 2] ), 'error in ex_width id={0}'.format(i)
         assert np.all(roidb[i]['ex_boxes'][:, 1] < roidb[i]['ex_boxes'][:, 3] ), 'error in ex_height id={0}'.format(i)
